@@ -6,8 +6,9 @@
   "scripts": {
     "build": "npm run clean && NODE_ENV=production ./node_modules/.bin/webpack -p --progress --config ./webpack.config.js",
     "clean": "rm -rf dist",
-    "dev": "NODE_ENV=development ./node_modules/.bin/webpack-dev-server --inline --hot",
-    "lint": "./node_modules/.bin/eslint app/"
+    "start": "NODE_ENV=development ./node_modules/.bin/webpack-dev-server --inline --hot",
+    "lint": "./node_modules/.bin/eslint src/",
+    "test": "npm run lint && ./node_modules/.bin/jest"
   },
   "main": "src/index.js",
   "repository": {
@@ -18,8 +19,9 @@
   "dependencies": {
     "google-map-react": "^0.22.3",
     "jquery": "^2.1.4",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "moment": "^2.17.1",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
     "react-redux": "^4.4.5",
     "react-router": "^2.8.1",
     "redux": "^3.6.0",
@@ -27,28 +29,41 @@
     "semantic-ui-css": "^2.1.8"
   },
   "devDependencies": {
-    "babel-core": "^6.3.17",
-    "babel-eslint": "^4.1.6",
-    "babel-loader": "^6.2.0",
-    "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
+    "babel-core": "^6.22.1",
+    "babel-eslint": "^7.1.1",
+    "babel-loader": "^6.2.10",
+    "babel-polyfill": "^6.22.0",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.22.0",
+    "babel-preset-stage-0": "^6.22.0",
+    "chai": "^3.5.0",
     "classnames": "^2.2.5",
     "css-loader": "^0.21.0",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^2.1.0",
-    "eslint-plugin-react": "^3.11.3",
-    "file-loader": "^0.8.4",
-    "jsx-loader": "^0.13.2",
-    "meaning-error": "^2.1.0",
+    "enzyme": "^2.7.1",
+    "eslint": "^3.15.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.9.0",
+    "fetch-mock": "^5.9.3",
+    "file-loader": "^0.10.0",
+    "jest": "^18.1.0",
     "node-sass": "^3.4.0",
+    "react-addons-test-utils": "^15.4.2",
     "react-ga": "^2.1.0",
+    "redux-mock-store": "^1.2.2",
     "sass-loader": "^3.0.0",
+    "sinon": "^1.17.7",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.6",
-    "webpack": "^1.12.2",
+    "webpack": "^1.14.0",
     "webpack-dev-server": "^1.12.1"
   },
-  "license": "MIT"
+  "license": "MIT",
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less|scss)$": "<rootDir>/spec/__mocks__/styleMock.js"
+    }
+  }
 }
