@@ -1,26 +1,37 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
+import React, { Component, PropTypes } from 'react'
+import { connect } from 'react-redux'
 
-import { fetchSocialList } from '../actions/social';
-import { strings } from '../config';
-import Image from '../components/image';
-import SocialItem from '../components/social-item';
-import './home.scss';
+import { fetchSocialList } from '../actions/social'
+import { strings } from '../config'
+import Icon from '../components/icon'
+import Image from '../components/image'
+import SocialItem from '../components/social-item'
+import './home.scss'
 
 
 class Home extends Component {
-  componentWillMount () {
-    this.props.dispatch(fetchSocialList());
+  static propTypes = {
+    dispatch: PropTypes.func,
+    socialList: PropTypes.object,
+  }
+
+  static defaultProps = {
+    dispatch: () => { },
+    socialList: {},
+  }
+
+  componentDidMount () {
+    this.props.dispatch(fetchSocialList())
   }
 
   renderMyImage () {
     return (
       <Image src={strings.app.gravatarUrl} />
-    );
+    )
   }
 
   renderSayMyName () {
-    return <h2 className="myName">{strings.app.appName}</h2>;
+    return <h2 className="myName">{strings.app.appName}</h2>
   }
 
   renderSayMyTitle () {
@@ -28,31 +39,29 @@ class Home extends Component {
       <div className="rubberBand animated myTitle">
         <h3>
           {strings.app.myTitle}
-          <i className="code icon"></i>
+          <Icon icon="code" />
         </h3>
       </div>
-    );
+    )
   }
 
   renderSocialList ({ items }) {
     return (
       <div className="social-list">
         {
-          items.map((item, index) => {
-            return (
-              <SocialItem
-                key={index}
-                item={item}
-              />
-            );
-          })
+          items.map(item => (
+            <SocialItem
+              key={item.icon}
+              item={item}
+            />
+          ))
         }
       </div>
-    );
+    )
   }
 
   render () {
-    const { socialList } = this.props;
+    const { socialList } = this.props
 
     return (
       <div className="page-home">
@@ -61,13 +70,11 @@ class Home extends Component {
         {this.renderSayMyTitle()}
         {!socialList.isFetching && this.renderSocialList(socialList)}
       </div>
-    );
+    )
   }
 }
 
 
-export default connect(state => {
-  return {
-    socialList: state.socialList,
-  };
-})(Home);
+export default connect(state => ({
+  socialList: state.socialList,
+}))(Home)
