@@ -1,9 +1,9 @@
 
 class ContactException {
   constructor (message) {
-    this.message = message;
-    this.name = 'ContactException';
+    this.message = message
+    this.name = 'ContactException'
   }
 }
 
-export default ContactException;
+export default ContactException
