@@ -1,8 +1,9 @@
-import React, { Component, PropTypes } from 'react';
-import classNames from 'classnames';
+import React, { Component, PropTypes } from 'react'
+import classNames from 'classnames'
 
-import FormField from './form-field';
-import { strings } from '../config';
+import FormField from './form-field'
+import Icon from './icon'
+import { strings } from '../config'
 
 
 class ContactForm extends Component {
@@ -10,10 +11,14 @@ class ContactForm extends Component {
     error: PropTypes.object,
     isSubmiting: PropTypes.bool.isRequired,
     onSubmit: PropTypes.func.isRequired,
-  };
+  }
+
+  static defaultProps = {
+    error: {},
+  }
 
   constructor (props) {
-    super(props);
+    super(props)
 
     this.state = {
       errorFields: {
@@ -24,89 +29,89 @@ class ContactForm extends Component {
         validation: false,
       },
       hasError: false,
-    };
+    }
 
-    this.onSubmit = this.onSubmit.bind(this);
+    this.onSubmit = this.onSubmit.bind(this)
   }
 
   onSubmit () {
-    if (this.props.isSubmiting) return;
+    if (this.props.isSubmiting) return
 
     const payload = {
-      name: this.refs.name.value,
-      email: this.refs.email.value,
-      subject: this.refs.subject.value,
-      message: this.refs.message.value,
-      validation: this.refs.validation.value,
-    };
+      name: this.name.value,
+      email: this.email.value,
+      subject: this.subject.value,
+      message: this.message.value,
+      validation: this.validation.value,
+    }
 
-    const { errorFields } = this.state;
-    const error = { ...errorFields };
+    const { errorFields } = this.state
+    const error = { ...errorFields }
     Object
       .keys(error)
-      .forEach(field => { error[field] = !payload[field]; });
+      .forEach((field) => { error[field] = !payload[field] })
 
-    if (parseInt(payload.validation, 10) !== 8) error.validation = true;
+    if (parseInt(payload.validation, 10) !== 8) error.validation = true
 
     const hasError = Object
       .keys(error)
       .filter(field => error[field])
-      .length > 0;
+      .length > 0
 
-    this.setState({ errorFields: error, hasError });
+    this.setState({ errorFields: error, hasError })
 
-    if (hasError) return;
+    if (hasError) return
 
-    this.props.onSubmit(payload);
+    this.props.onSubmit(payload)
   }
 
   renderErrorMessage (error) {
-    if (!error) return null;
+    if (!error) return null
 
-    let errorMessage = strings.contact.form.errorMessage;
+    let errorMessage = strings.contact.form.errorMessage
     if (error.message) {
-      errorMessage = error.message;
+      errorMessage = error.message
     }
 
     return (
       <div className="ui icon error message">
-        <i className="warning sign icon"></i>
+        <Icon icon="warning sign" />
         <div className="content">
           <div className="header">
             {strings.contact.form.errorHeader}
           </div>
-          <p dangerouslySetInnerHTML={{ __html: errorMessage }}></p>
+          <p>{errorMessage}</p>
         </div>
       </div>
-    );
+    )
   }
 
   render () {
-    const { error, isSubmiting } = this.props;
-    const { errorFields, hasError } = this.state;
+    const { error, isSubmiting } = this.props
+    const { errorFields, hasError } = this.state
     const formCssClass = classNames(
       'contact-form-component ui form',
-      { 'error': hasError || error }
-    );
+      { error: hasError || error },
+    )
     const submitCssClass = classNames(
       'ui green submit button',
-      { 'loading': isSubmiting }
-    );
-    const formStrings = strings.contact.form;
+      { loading: isSubmiting },
+    )
+    const formStrings = strings.contact.form
 
     return (
       <div className={formCssClass}>
         <div className="two fields">
           <FormField label={formStrings.name.label} hasError={errorFields.name}>
             <input
-              ref="name"
+              ref={(name) => { this.name = name }}
               type="text"
               placeholder={formStrings.name.placeholder}
             />
           </FormField>
           <FormField label={formStrings.email.label} hasError={errorFields.email}>
             <input
-              ref="email"
+              ref={(email) => { this.email = email }}
               type="email"
               placeholder={formStrings.email.placeholder}
             />
@@ -114,28 +119,28 @@ class ContactForm extends Component {
         </div>
         <FormField label={formStrings.subject.label} hasError={errorFields.subject}>
           <input
-            ref="subject"
+            ref={(subject) => { this.subject = subject }}
             type="text"
             placeholder={formStrings.subject.placeholder}
           />
         </FormField>
         <FormField label={formStrings.message.label} hasError={errorFields.message}>
-          <textarea ref="message" rows="4"></textarea>
+          <textarea ref={(message) => { this.message = message }} rows="4" />
         </FormField>
         <FormField label={formStrings.validation.label} hasError={errorFields.validation}>
           <input
-            ref="validation"
+            ref={(validation) => { this.validation = validation }}
             type="text"
             placeholder={formStrings.validation.placeholder}
           />
         </FormField>
         {this.renderErrorMessage(error)}
-        <div className={submitCssClass} onClick={this.onSubmit}>
+        <button className={submitCssClass} onClick={this.onSubmit}>
           {formStrings.send}
-        </div>
+        </button>
       </div>
-    );
+    )
   }
 }
 
-export default ContactForm;
+export default ContactForm
