@@ -1,26 +1,21 @@
-import React, { Component, PropTypes } from 'react';
+import React, { PropTypes } from 'react'
 
+import Icon from './icon'
 
-class SocialItem extends Component {
-  static propTypes = {
-    item: PropTypes.object.isRequired,
-  };
+const SocialItem = ({ item }) => (
+  <a
+    data-content={item.url}
+    className={`ui circular icon button ${item.className || item.icon}`}
+    href={item.url}
+    target="_blank"
+    rel="noopener noreferrer"
+  >
+    <Icon icon={item.icon} />
+  </a>
+)
 
-  render () {
-    const { item } = this.props;
-
-    return (
-      <a
-        ref="socialItem"
-        data-content={item.url}
-        className={`ui circular icon button ${item.className || item.icon}`}
-        href={item.url}
-        target="_blank"
-      >
-        <i className={`${item.icon} icon`}></i>
-      </a>
-    );
-  }
+SocialItem.propTypes = {
+  item: PropTypes.object.isRequired,
 }
 
-export default SocialItem;
+export default SocialItem
