@@ -1,14 +1,14 @@
-import { combineReducers } from 'redux';
+import { combineReducers } from 'redux'
 
 
-import sendContact from './send-contact';
-import socialList from './social-list';
+import sendContact from './send-contact'
+import socialList from './social-list'
 
 
 const rootReducer = combineReducers({
   sendContact,
   socialList,
-});
+})
 
 
-export default rootReducer;
+export default rootReducer
