@@ -1,14 +1,14 @@
-import React from 'react';
-import { Route, IndexRoute } from 'react-router';
+import React from 'react'
+import { Route, IndexRoute } from 'react-router'
 
-import App from './containers/app';
-import Home from './containers/home';
-import Contact from './containers/contact';
+import App from './containers/app'
+import Home from './containers/home'
+import Contact from './containers/contact'
 
 
 export default (
   <Route path="/" component={App}>
-    <IndexRoute component={Home}/>
+    <IndexRoute component={Home} />
     <Route path="/contact" component={Contact} />
   </Route>
-);
+)
